Implements OAuth using Ueberauth
This code uses Ueberauth dynamically to make providers able to be stored
in database. Ueberauth specific parts are split in separate functions.
The only non dynamic part is the OAuth Client ID and secret for GitHub.
All GitHub providers share the same values passed from the environment.

As this involves connecting with external APIs, it's difficult to find
relevant tests to make and only minimal tests are included.